Rename P_CacheVol.h to StripeSM.h
