+++ src/iocore/cache/Stripe.cc
@@ -23,7 +23,7 @@
 
 #include "P_CacheDoc.h"
 #include "P_CacheInternal.h"
-#include "P_CacheVol.h"
+#include "StripeSM.h"
 
 #include "tscore/ink_assert.h"
 #include "tscore/ink_memory.h"
