+++ src/iocore/cache/Stripe.h
@@ -43,7 +43,7 @@
 #define ROUND_TO_SECTOR(_p, _x)  INK_ALIGN((_x), _p->sector_size)
 #define ROUND_TO(_x, _y)         INK_ALIGN((_x), (_y))
 
-// This is defined here so CacheVC can avoid including P_CacheVol.h.
+// This is defined here so CacheVC can avoid including StripeSM.h.
 #define RECOVERY_SIZE EVACUATION_SIZE // 8MB
 
 struct CacheVol {
