+++ src/iocore/cache/StripeSM.h
